Update the Taken time for photos
When i uploaded all my Whatsapp photos/videos on 'Google Photos', i noticed that the photos did not have proper Exif data due to which the photos were placed in the wrong timestamp i.e. the current timestamp. It is easy to fix this:
For eg: The Whatsapp Photo/Video Format is [IMG-20150110-WA0018.mp4] or [IMG-20150110-WA0018.jpg] so we get the year, month, date from the file name itself. This is exactly what i am doing after which i am updating the exif data using piexif.